AP_GPS_SBP: RTK Support for Swift Binary Protocol GPSes
This patch splits the SBP GPS driver into a two-state state machine: Single Point Positioning mode and RTK mode.
The driver starts in Single Point Positionin mode, and if both a home LLH (Lat Lon Height) is captured and RTK-calculated baseline messages are received, it switches into RTK mode.

In RTK mode, the LLH position reported by the driver is calculated by adding the RTK-accurate baseline to the SPP-accurate home position. Thus, for flight relative to the home position, it has centimeter-accurate positioning. In terms of absolute location, it is still SPP accurate.